requirements.txt
absl-py==0.8.1
appnope==0.1.0
astor==0.8.0
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
Bottleneck==1.3.1
certifi==2019.11.28
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
future==0.18.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
imageio==2.6.1
importlib-metadata==1.5.0
inflect==4.1.0
innvestigate==1.0.8
ipdb==0.12.2
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jaraco.itertools==5.0.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lime==0.1.1.36
Markdown==3.1.1
MarkupSafe==1.1.1
matlabengineforpython===R2018a
matplotlib==3.1.1
mistune==0.8.4
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
notebook==6.0.3
numpy==1.17.2
opt-einsum==3.1.0
packaging==19.2
pandas==0.25.1
pandocfilters==1.4.2
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.0
pluggy==0.13.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.10.0
ptyprocess==0.6.0
py==1.8.0
Pygments==2.5.2
pyparsing==2.4.2
PyQt5==5.14.1
PyQt5-sip==12.7.1
pyrsistent==0.15.7
pytest==5.2.1
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==18.1.1
qtconsole==4.6.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
tensorboard==1.12.2
tensorflow==1.12.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tk==0.1.0
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==1.2.0
zipp==2.1.0
